Prevented new position from being negative
It turns out that fmod returns can sometimes return a negative value
which is congruent to the 'correct' positive value. This was causing
issues as I was previously trying to set the new position to a negative
value, which, really, I wanted to be a large, positive value (i.e. I
wanted the boid to 'loop around' the board). This should fix that bug.