Wrote function 'positive_fmod'
This was necessary because fmod's ability to return negative numbers was
causing problems in lots of places in my program. I used the logic laid
out in a StackOverflow answer
(http://stackoverflow.com/a/29871193/6914552) to write a function which
returns the smallest non-negative number congruent to that which is
returned by fmod.