Fixed bug with boundaries in cohesion vector
The function calculating the new cohesion vector was previously unable
to deal with the periodic boundary conditions imposed on the board. For
example, if a boid at (1,1) had two other boids in its neighbourhood,
positioned at (1,3) and (99, 2) (with a max_x of 100) then the centroid
was being computed as being at (33.7, 2), whereas it clearly it located
at (1/3, 2). This fixes that behaviour.