Constrain individual avoidance vectors to 1
To see why this change is necessary, consider two pairs of boids:

    Pair 1: A pair of boids with positions (0,0) and (2,2)
    Pair 2: A pair of boids with positions (1,1) and (99,99) (assuming the
            board has width 100)

Notice that the shortest path between the pair in Pair 2 crosses a
boundary.

Previously, the avoidance vector calculation would return a vector in
the correct direction, but not care about the magnitude of this vector.
This meant that the displacement vectors between the boids in these two
pairs will differ considerably.  This clearly is an issue, as, in terms
of the computation of the avoidance vector, these two pairs of boids are
the same. Therefore, to remedy this issue, I am now constraining all
avoidance vectors between pairs of boids to unit length.