Stopped edge case glitch when computing centroid
I need to use one of two methods when computing the centroid of a
neighbourhood. The simpler one is for when the neighbourhood does not
cross a boundary, and the more complex one is for when the neighbourhood
*does* cross a boundary. I spent a lot of time trying to use just on or
the other method, but I have finally concluded that this is not
possible.