Fixed vector magnitude when crossing boundaries
Before, say I was wanting the shortest displacement vector from (99, 1)
to (1, 50), I was ending up with a vector of (98, 49). This is an issue
because it makes it seem like we need to travel further in the x
direction than we do in the y direction. In fact, the opposite is true.
This change should fix this bug, so that the vector computed now would
be (2, 49).