Changed how new velocity calculated
Previously, if the boid was in the 'danger zone', it would only use the
avoidance velocity. This led to flocks dispersing too easily. Now, if
the boid is not in a danger zone, it uses all steers except avoidance,
and if it is in the danger zone, it uses all steers. This leads to the
best example of flocking which I have yet seen.